# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tanabata API",
    "contact": {
      "name": "Tanabata maintainers",
      "url": "https://github.com/gongahkia/tanabata/issues"
    },
    "license": {
      "name": "MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/v1/works": {
      "get": {
        "operationId": "listWorks",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "artist_id", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "maximum": 10000 } }
        ],
        "responses": {
          "200": {
            "description": "Works",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Work" } },
                    "meta": { "$ref": "#/components/schemas/ListMeta" },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "default": {
            "$ref": "#/components/responses/ErrorDefault"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        }
      }
    },
...
    "/v1/recordings": {
      "get": {
        "operationId": "listRecordings",
        "parameters": [
          { "name": "artist_id", "in": "query", "schema": { "type": "string" } },
          { "name": "work_id", "in": "query", "schema": { "type": "string" } },
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "maximum": 10000 } }
        ],
        "responses": {
          "200": {
            "description": "Recordings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Recording" } },
                    "meta": { "$ref": "#/components/schemas/ListMeta" },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "default": {
            "$ref": "#/components/responses/ErrorDefault"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        }
      }
    },
...
    "/v1/claims": {
      "get": {
        "operationId": "listClaims",
        "parameters": [
          { "name": "kind", "in": "query", "schema": { "type": "string", "enum": ["attribution", "sample", "credit", "cover", "performance"] } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "maximum": 10000 } }
        ],
        "responses": {
          "200": {
            "description": "Unified claim feed across the catalog",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Claim" } },
                    "meta": { "$ref": "#/components/schemas/ListMeta" },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "default": {
            "$ref": "#/components/responses/ErrorDefault"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Pagination": {
        "type": "object",
        "properties": {
          "limit": { "type": "integer" },
          "offset": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
      "ListMeta": {
        "type": "object",
        "properties": {
          "snapshot_version": { "type": "string" },
          "active_providers": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "Source": {
        "type": "object",
        "properties": {
          "source_id": { "type": "string" },
          "provider": { "type": "string" },
          "url": { "type": "string" },
          "title": { "type": "string" },
          "publisher": { "type": "string" },
          "license": { "type": "string" },
          "retrieved_at": { "type": "string" }
        }
      },
...
      "Work": {
        "type": "object",
        "properties": {
          "work_id": { "type": "string" },
          "mbid": { "type": "string" },
          "title": { "type": "string" },
          "iswc": { "type": "string" },
          "language": { "type": "string" },
          "created_year": { "type": "string" },
          "primary_artist_id": { "type": "string" },
          "primary_artist_name": { "type": "string" },
          "notes": { "type": "string" },
          "recording_count": { "type": "integer" },
          "credit_count": { "type": "integer" },
          "disputed_credit_count": { "type": "integer" },
          "tags": { "type": "array", "items": { "type": "string" } }
        }
      },
      "Recording": {
        "type": "object",
        "properties": {
          "recording_id": { "type": "string" },
          "mbid": { "type": "string" },
          "work_id": { "type": "string" },
          "work_title": { "type": "string" },
          "artist_id": { "type": "string" },
          "artist_name": { "type": "string" },
          "title": { "type": "string" },
          "duration_ms": { "type": "integer" },
          "released_year": { "type": "string" },
          "release_id": { "type": "string" },
          "isrc": { "type": "string" },
          "is_original": { "type": "boolean" },
          "notes": { "type": "string" },
          "sample_outgoing_count": { "type": "integer" },
          "sample_incoming_count": { "type": "integer" }
        }
      },
...
      "Claim": {
        "type": "object",
        "properties": {
          "claim_id": { "type": "string" },
          "kind": { "type": "string" },
          "subject_type": { "type": "string" },
          "subject_id": { "type": "string" },
          "object_type": { "type": "string" },
          "object_id": { "type": "string" },
          "relation": { "type": "string" },
          "status": { "type": "string" },
          "confidence_score": { "type": "number" },
          "provider_origin": { "type": "string" },
          "source_id": { "type": "string" },
          "asserted_at": { "type": "string" },
          "last_verified_at": { "type": "string" },
          "notes": { "type": "string" },
          "source": { "$ref": "#/components/schemas/Source" },
          "supporting_evidence_count": { "type": "integer" },
          "refuting_evidence_count": { "type": "integer" },
          "evidence": { "type": "array", "items": { "$ref": "#/components/schemas/ClaimEvidence" } }
        }
      },
...
      "ClaimEvidence": {
        "type": "object",
        "properties": {
          "evidence_id": { "type": "string" },
          "claim_id": { "type": "string" },
          "supports": { "type": "boolean" },
          "source_id": { "type": "string" },
          "excerpt": { "type": "string" },
          "source_url": { "type": "string" },
          "archived_url": { "type": "string" },
          "evidence_kind": { "type": "string" },
          "weight": { "type": "number" },
          "recorded_at": { "type": "string" },
          "source": { "$ref": "#/components/schemas/Source" }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "code"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string" },
          "code": { "type": "string" },
          "details": { "type": "object", "additionalProperties": true }
        }
      }
    },
    "securitySchemes": {
      "PublicNoAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-No-Auth-Required",
        "description": "Documentation-only marker: Tanabata API endpoints are public and require no authentication."
      }
    },
    "responses": {
      "Error400": {
        "description": "Bad Request",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" },
            "examples": {
              "missingQuery": {
                "value": {
                  "type": "https://tanabata.dev/errors/missing_query",
                  "title": "q is required",
                  "status": 400,
                  "detail": "q is required",
                  "instance": "req_01HX0000000000000000000000",
                  "code": "missing_query"
                }
              }
            }
          }
        }
      },
      "Error404": {
        "description": "Not Found",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" },
            "examples": {
              "artistNotFound": {
                "value": {
                  "type": "https://tanabata.dev/errors/artist_not_found",
                  "title": "artist not found",
                  "status": 404,
                  "detail": "artist not found",
                  "instance": "req_01HX0000000000000000000001",
                  "code": "artist_not_found"
                }
              }
            }
          }
        }
      },
      "Error409": {
        "description": "Conflict",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" },
            "examples": {
              "conflict": {
                "value": {
                  "type": "https://tanabata.dev/errors/conflict",
                  "title": "resource conflict",
                  "status": 409,
                  "detail": "resource conflict",
                  "instance": "req_01HX0000000000000000000002",
                  "code": "conflict"
                }
              }
            }
          }
        }
      },
      "Error500": {
        "description": "Internal Server Error",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" },
            "examples": {
              "internalError": {
                "value": {
                  "type": "https://tanabata.dev/errors/internal_error",
                  "title": "unexpected server error",
                  "status": 500,
                  "detail": "unexpected server error",
                  "instance": "req_01HX0000000000000000000003",
                  "code": "internal_error"
                }
              }
            }
          }
        }
      },
      "ErrorDefault": {
        "description": "Error response",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" },
            "examples": {
              "defaultError": {
                "value": {
                  "type": "https://tanabata.dev/errors/internal_error",
                  "title": "unexpected server error",
                  "status": 500,
                  "detail": "unexpected server error",
                  "instance": "req_01HX0000000000000000000004",
                  "code": "internal_error"
                }
              }
            }
          }
        }
      },
      "Error429": {
        "description": "Too Many Requests",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": { "type": "string" }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" },
            "examples": {
              "rateLimited": {
                "value": {
                  "type": "https://tanabata.dev/errors/rate_limited",
                  "title": "too many requests",
                  "status": 429,
                  "detail": "too many requests",
                  "instance": "req_01HX0000000000000000000005",
                  "code": "rate_limited"
                }
              }
            }
          }
        }
      }
    }
  }
}